Add SSL request support
